Fully Recovered

+++ src/main/java/com/mycompany/reservationsystem/dto/ReservationCustomerDTO.java
@@ -12,7 +12,7 @@ public class ReservationCustomerDTO {
     private String status;
     private String reference;
     private LocalDate date;
-    private int revenue;
+    private Double revenue;
     private LocalTime reservationPendingtime;
     private LocalTime reservationConfirmtime;
     private LocalTime reservationCancelledtime;
@@ -27,7 +27,7 @@ public class ReservationCustomerDTO {
 
     // Constructor
     public ReservationCustomerDTO(Long reservationId, int pax, String prefer, String status, String reference,
-                                  LocalDate date, int revenue, LocalTime reservationPendingtime,
+                                  LocalDate date, Double revenue, LocalTime reservationPendingtime,
                                   LocalTime reservationConfirmtime, LocalTime reservationCancelledtime,
                                   LocalTime reservationSeatedtime, LocalTime reservationCompletetime,
                                   Long customerId, String customerName, String customerPhone, String customerEmail) {
@@ -60,7 +60,7 @@ public class ReservationCustomerDTO {
     public String getStatus() { return status; }
     public String getReference() { return reference; }
     public LocalDate getDate() { return date; }
-    public int getRevenue() { return revenue; }
+    public Double getRevenue() { return revenue; }
     public LocalTime getReservationPendingtime() { return reservationPendingtime; }
     public LocalTime getReservationConfirmtime() { return reservationConfirmtime; }
     public LocalTime getReservationCancelledtime() { return reservationCancelledtime; }
